Fix 3 issues when search paths span multiple filesystems:
1. Duplicates were only found within 1 of the filesystems.
This is issue #3 in the issue list and was reported
and fixed by Robert Seczkowski.
Note this bug only affects systems with findutils >= 4.2

2. Disk head seeks were very inefficient when there
were overlapping inode numbers between files of the
same size on different filesystems (which is very likely).
We now sort by device number and then by inode.
Note will still be an issue on systems with findutils < 4.2

3. When merging duplicates using hardlinks,
and those duplicates are on different filesystems,
use symlinks rather than just deleting the files.


git-svn-id: https://fslint.googlecode.com/svn/trunk@11 494acc53-a219-0410-adff-8f8ff343f446