Set mouse pointer to "busy interactive" during searches. This makes it more obvious that work is progressing and only parts of the GUI will be responsive.
This was much harder than it sounds due to gtk not
providing a nice interface to set a cursor for an app,
or even a widget. Also gtk hasn't direct support for
the standard "busy interactive" cursor that mozilla uses
(left_ptr_watch), so I needed to copy the actual cursor
bits that mozilla uses!

This fixes fedora bug 234874



git-svn-id: https://fslint.googlecode.com/svn/trunk@60 494acc53-a219-0410-adff-8f8ff343f446