Change the duplicate merging algorithm from:
  unlink(dup_name)
  link(dup_name, data_name)

to:

  link(tmp_name, data_name)
  rename(tmp_name, dup_name)
  unlink(tmp_name) #in case error condition

This is essentially so that the unlink
is only done at the end.
Previously if the "keep" file was removed,
before the merge process was run, all files
in the duplicate group would be deleted.
Also if the filesystem didn't support links,
the duplicate files names that were to be linked
would be loast.



git-svn-id: https://fslint.googlecode.com/svn/trunk@62 494acc53-a219-0410-adff-8f8ff343f446