don't auto remove duplicate groups when renaming
If we rename a file, then we remove it from the
list as implicitly we would want to keep it.
But if there was just one other duplicate file in the group,
then that would also have been removed from display,
even though the user may want to delete it etc.

Addresses part of issue #119